+++ integration_test/utils/connect_to_pod.dart
@@ -34,8 +34,9 @@ import 'package:flutter_test/flutter_test.dart';
 import 'package:integration_test/integration_test.dart';
 
 import 'package:healthpod/main.dart' as app;
-import 'package:healthpod/utils/fetch_web_id.dart';
 import 'package:healthpod/utils/create_solid_login.dart';
+import 'package:healthpod/utils/fetch_web_id.dart';
+
 
 /// Integration tests for Solid POD connection functionality in HealthPod app.
 ///
